+++ kernel_tuner/backends/nvcuda.py
@@ -2,7 +2,7 @@
 import numpy as np
 
 from kernel_tuner.backends.backend import GPUBackend
-from kernel_tuner.observers import BenchmarkObserver
+from kernel_tuner.observers.observer import BenchmarkObserver
 from kernel_tuner.util import SkippableFailure
 
 #embedded in try block to be able to generate documentation
